Split libraries into multiple modules
