+++ scriptinator-docs/src/main/asciidoc/libraries/_snippets/functions/HTTP/HttpClient/head_options/parameters.adoc
@@ -14,5 +14,4 @@
 // limitations under the License.
 //
 
-:signature: get({url, contentType, body, headers})
-:description: Get data from the server using a `GET` request.
+include::../get_options/parameters.adoc[]
