+++ scriptinator-docs/src/main/asciidoc/libraries/http.adoc
@@ -95,3 +95,8 @@ include::_snippets/function.adoc[]
 :function_id: request_options
 include::_snippets/function.adoc[]
 
+== HttpResponse
+:type_name: HttpResponse
+
+== HttpResponseBody
+:type_name: HttpResponseBody
