+++ scriptinator-lib-assert/pom.xml
@@ -16,8 +16,8 @@
     limitations under the License.
 
 -->
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xmlns="http://maven.apache.org/POM/4.0.0"
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
@@ -27,19 +27,13 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>scriptinator-docs</artifactId>
+    <artifactId>scriptinator-lib-assert</artifactId>
 
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.asciidoctor</groupId>
-                <artifactId>asciidoctor-maven-plugin</artifactId>
-            </plugin>
-            <plugin>
-                <groupId>com.spotify</groupId>
-                <artifactId>dockerfile-maven-plugin</artifactId>
-            </plugin>
-        </plugins>
-    </build>
+    <dependencies>
+        <dependency>
+            <groupId>io.chapp.scriptinator</groupId>
+            <artifactId>scriptinator-lib-core</artifactId>
+        </dependency>
+    </dependencies>
 
 </project>
