+++ scriptinator-lib-core/src/main/java/io/chapp/scriptinator/libraries/core/ClosableContext.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.chapp.scriptinator;
+package io.chapp.scriptinator.libraries.core;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
