+++ scriptinator-lib-core/src/main/java/io/chapp/scriptinator/libraries/core/DataValue.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.chapp.scriptinator.libraries;
+package io.chapp.scriptinator.libraries.core;
 
-import org.springframework.util.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import java.util.HashMap;
 import java.util.Map;
