+++ scriptinator-lib-core/src/main/java/io/chapp/scriptinator/libraries/core/EncodeUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.chapp.scriptinator.libraries;
+package io.chapp.scriptinator.libraries.core;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
