+++ scriptinator-lib-http/src/main/java/io/chapp/scriptinator/libraries/http/BasicAuthentication.java
@@ -15,23 +15,33 @@
  */
 package io.chapp.scriptinator.libraries.http;
 
-import io.chapp.scriptinator.libraries.DataValue;
+import io.chapp.scriptinator.libraries.core.DataValue;
 
 import java.util.Map;
 
-public class BearerAuthentication extends DataValue {
-    private String token;
+public class BasicAuthentication extends DataValue {
+    private String username;
+    private String password;
 
-    public String getToken() {
-        return token;
+    public String getUsername() {
+        return username;
     }
 
-    public void setToken(String token) {
-        this.token = token;
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
     }
 
     @Override
     protected void inspect(Map<String, Object> target) {
-        target.put("token", token == null ? "<missing>" : "****");
+        target.put("username", username);
+        target.put("password", password == null ? "<missing>" : "****");
     }
 }
