+++ scriptinator-lib-http/src/main/java/io/chapp/scriptinator/libraries/http/HttpResponse.java
@@ -15,7 +15,7 @@
  */
 package io.chapp.scriptinator.libraries.http;
 
-import io.chapp.scriptinator.libraries.DataValue;
+import io.chapp.scriptinator.libraries.core.DataValue;
 import okhttp3.Response;
 import okhttp3.ResponseBody;
