+++ scriptinator-lib-http/src/main/java/io/chapp/scriptinator/libraries/http/RequestOptions.java
@@ -15,7 +15,8 @@
  */
 package io.chapp.scriptinator.libraries.http;
 
-import io.chapp.scriptinator.libraries.DataValue;
+
+import io.chapp.scriptinator.libraries.core.DataValue;
 
 import java.util.Map;
